backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
rembg==2.0.50
Pillow==10.2.0
httpx==0.26.0
email-validator==2.1.0.post1
pytest==8.0.2